Move partition utility methods to Partitions class
