+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/HivePartitionManager.java
@@ -38,7 +38,7 @@ import java.util.function.Predicate;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.collect.ImmutableList.toImmutableList;
-import static io.trino.metastore.Partition.unescapePathName;
+import static io.trino.metastore.Partitions.unescapePathName;
 import static io.trino.plugin.hive.metastore.MetastoreUtil.computePartitionKeyFilter;
 import static io.trino.plugin.hive.metastore.MetastoreUtil.toPartitionName;
 import static io.trino.plugin.hive.util.HiveBucketing.getHiveBucketFilter;
