+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/metastore/HivePartitionName.java
@@ -23,7 +23,7 @@ import java.util.Objects;
 import java.util.Optional;
 
 import static com.google.common.base.MoreObjects.toStringHelper;
-import static io.trino.metastore.Partition.toPartitionValues;
+import static io.trino.metastore.Partitions.toPartitionValues;
 import static java.util.Objects.requireNonNull;
 
 @Immutable
