+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/projection/PartitionProjection.java
@@ -33,9 +33,9 @@ import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
 import static com.google.common.collect.Sets.cartesianProduct;
-import static io.trino.metastore.Partition.toPartitionValues;
+import static io.trino.metastore.Partitions.escapePathName;
+import static io.trino.metastore.Partitions.toPartitionValues;
 import static io.trino.plugin.hive.projection.InvalidProjectionException.invalidProjectionMessage;
-import static io.trino.plugin.hive.util.HiveUtil.escapePathName;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
