+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/util/TestHiveUtil.java
@@ -13,20 +13,11 @@
  */
 package io.trino.plugin.hive.util;
 
-import io.trino.metastore.Partition;
-import org.apache.hadoop.hive.common.FileUtils;
-import org.apache.hadoop.hive.metastore.Warehouse;
-import org.apache.hadoop.hive.metastore.api.MetaException;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 import org.joda.time.format.DateTimeFormat;
 import org.junit.jupiter.api.Test;
 
-import java.util.AbstractList;
-import java.util.ArrayList;
-import java.util.List;
-
-import static io.trino.metastore.Partition.toPartitionValues;
 import static io.trino.plugin.hive.util.HiveUtil.escapeSchemaName;
 import static io.trino.plugin.hive.util.HiveUtil.escapeTableName;
 import static io.trino.plugin.hive.util.HiveUtil.parseHiveTimestamp;
@@ -45,43 +36,6 @@ public class TestHiveUtil
         assertThat(parse(time, "yyyy-MM-dd HH:mm:ss.SSS")).isEqualTo(unixTime(time, 3));
         assertThat(parse(time, "yyyy-MM-dd HH:mm:ss.SSSSSSS")).isEqualTo(unixTime(time, 6));
         assertThat(parse(time, "yyyy-MM-dd HH:mm:ss.SSSSSSSSS")).isEqualTo(unixTime(time, 7));
-    }
-
-    @Test
-    public void testToPartitionValues()
-            throws MetaException
-    {
-        assertToPartitionValues("ds=2015-12-30/event_type=QueryCompletion");
-        assertToPartitionValues("ds=2015-12-30");
-        assertToPartitionValues("a=1/b=2/c=3");
-        assertToPartitionValues("a=1");
-        assertToPartitionValues("pk=!@%23$%25%5E&%2A()%2F%3D");
-        assertToPartitionValues("pk=__HIVE_DEFAULT_PARTITION__");
-    }
-
-    @Test
-    public void testUnescapePathName()
-    {
-        assertUnescapePathName("", "");
-        assertUnescapePathName("x", "x");
-        assertUnescapePathName("abc", "abc");
-        assertUnescapePathName("abc%", "abc%");
-        assertUnescapePathName("%", "%");
-        assertUnescapePathName("%41", "A");
-        assertUnescapePathName("%41%x", "A%x");
-        assertUnescapePathName("%41%xxZ", "A%xxZ");
-        assertUnescapePathName("%41%%Z", "A%%Z");
-        assertUnescapePathName("%41%25%25Z", "A%%Z");
-        assertUnescapePathName("abc%41%42%43", "abcABC");
-        assertUnescapePathName("abc%3Axyz", "abc:xyz");
-        assertUnescapePathName("abc%3axyz", "abc:xyz");
-        assertUnescapePathName("abc%BBxyz", "abc\u00BBxyz");
-    }
-
-    private static void assertUnescapePathName(String value, String expected)
-    {
-        assertThat(FileUtils.unescapePathName(value)).isEqualTo(expected);
-        assertThat(Partition.unescapePathName(value)).isEqualTo(expected);
     }
 
     @Test
@@ -106,56 +60,6 @@ public class TestHiveUtil
                 .hasMessage("The provided tableName cannot be null or empty");
         assertThat(escapeTableName("../table1")).isEqualTo("..%2Ftable1");
         assertThat(escapeTableName("../../table1")).isEqualTo("..%2F..%2Ftable1");
-    }
-
-    @Test
-    public void testEscapePathName()
-    {
-        assertEscapePathName(null, "__HIVE_DEFAULT_PARTITION__");
-        assertEscapePathName("", "__HIVE_DEFAULT_PARTITION__");
-        assertEscapePathName("x", "x");
-        assertEscapePathName("abc", "abc");
-        assertEscapePathName("%", "%25");
-        assertEscapePathName("A", "A");
-        assertEscapePathName("A%x", "A%25x");
-        assertEscapePathName("A%xxZ", "A%25xxZ");
-        assertEscapePathName("A%%Z", "A%25%25Z");
-        assertEscapePathName("abcABC", "abcABC");
-        assertEscapePathName("abc:xyz", "abc%3Axyz");
-        assertEscapePathName("abc\u00BBxyz", "abc\u00BBxyz");
-        assertEscapePathName("\u0000\t\b\r\n\u001F", "%00%09%08%0D%0A%1F");
-        assertEscapePathName("#%^&*=[]{\\:'\"/?", "%23%25%5E&%2A%3D%5B%5D%7B%5C%3A%27%22%2F%3F");
-        assertEscapePathName("~`!@$()-_+}|;,.<>", "~`!@$()-_+}|;,.<>");
-    }
-
-    private static void assertEscapePathName(String value, String expected)
-    {
-        assertThat(FileUtils.escapePathName(value)).isEqualTo(expected);
-        assertThat(HiveUtil.escapePathName(value)).isEqualTo(expected);
-    }
-
-    @Test
-    public void testMakePartName()
-    {
-        assertMakePartName(List.of("abc"), List.of("xyz"), "abc=xyz");
-        assertMakePartName(List.of("abc:qqq"), List.of("xyz/yyy=zzz"), "abc%3Aqqq=xyz%2Fyyy%3Dzzz");
-        assertMakePartName(List.of("abc", "def", "xyz"), List.of("qqq", "rrr", "sss"), "abc=qqq/def=rrr/xyz=sss");
-    }
-
-    private static void assertMakePartName(List<String> columns, List<String> values, String expected)
-    {
-        assertThat(FileUtils.makePartName(columns, values)).isEqualTo(expected);
-        assertThat(HiveUtil.makePartName(columns, values)).isEqualTo(expected);
-    }
-
-    private static void assertToPartitionValues(String partitionName)
-            throws MetaException
-    {
-        List<String> actual = toPartitionValues(partitionName);
-        AbstractList<String> expected = new ArrayList<>();
-        actual.forEach(s -> expected.add(null));
-        Warehouse.makeValsFromName(partitionName, expected);
-        assertThat(actual).isEqualTo(expected);
     }
 
     private static long parse(DateTime time, String pattern)
